Add basic validation
This has the effect of disabling the paypal buttons until basic validate passes. The goal
here is to prevent the scenario where someone clicks on the paypal button, authorizes the payment
and then has to re-authorize due to the form submission not being accepted. We REALLY don't want
to take their authorization until the form is correct. Note that there is
an api we could use (ContributionPage.validate ['action' => 'create', 'id' => x]

This attempt https://github.com/eileenmcnaughton/nz.co.fuzion.omnipaymultiprocessor/pull/98/ was against
the previous version of paypal smartbuttons but it demonstrates the api

I couldn't find any evidence of being able to listen to when a form is valid - posts like this
https://github.com/jquery-validation/jquery-validation/issues/1996
suggest the method I used.

The easiest way to test this is by submitting the form as an anonymous user without
entering an email - you should not be able to launch the paypal form and there should be red writing next to
the field.

Paypal documentation on it

https://developer.paypal.com/docs/checkout/integration-features/validation/?mark=validat#synchronous-validation

Note the async mode is BAD - it launches the modal & does the check while loading. Avoid. Perhaps we could only render
the paypal button when valid - like they suggest - but then switching back & forth with valid & invalid gets hard.

Am on the fence about hiding
Unfortunately without a core patch you will ALSO get an alert. I am proposing a core patch along side this
which will suppress it in core. Really I think it would be best to put the message down by the paypal button
(possibly hiding it) but because of the need to call it at the start it needs to be less nasty.