Fix SagePay to correctly confirm the transaction.
This allows for the possiblity of the trxn_id to be saved before redirect, if available
(which is probably only true for SagePay at the moment) and for that to be reloaded
when completing the notification. If it is an array (i.e json encoded array)
then the parameters (if valid) are set on the response.
'confirm' is called if the method exists.

Note that the call to confirm has been moved to before completetransaction
as a successful confirm is necessary before the transaction is actually
complete. The move to put it in redirectOrExit earlier probably
reflects the fact that necessary output is done in this
function. I don't think doing it earlier is a problem
(for this or Mercanet which somewhat copied the pattern) but
if it is we can use ob_start, ob_get_contents etc