Sagepay recurring support
This works but there are a few outstanding considerations

1) Sagepay uses actions that are inconsistent. I have gone with declaring this as metadata
as the 'most' transparent metadata - ie each piece of metadata should not
have a bunch of interwoved assumptions like 'if it's this
action then it's also this action and card is handled this way' but
I hope we can do some upstream work
to make it less ad hoc  - https://github.com/thephpleague/omnipay-sagepay/issues/157

2) I'm a bit on the fence about the approach of creating a token only when it is recurring
and still using transaction data from the contribution.trxn_id. I think overall I prefer
to always create a token since any contribution could be used for a token and not
to save transaction data (over and above the trxn_id) in the contributon.trxn_id
but given I had written it this way I have not preferred that enough to re-write it. Test
cover should facilitate future changes (more or less)