Add warning regarding known gaps in meeting the contract if  property bag is used
Core does not use property bag - and there are no known instances of
people using property bag with Omnipay but if they did it
would fail to create recurrings as of 5.69 because it
does not pass our is_recur. Omnipay is not the only processor
that requires is_recur.

Note in the past Omnipay had a cast to PropertyBag but I reverted that
after hitting regressions in Omnipay caused by property bag

As far as I know there are no runtime benefits of using Omnipay